"Delete Section api"

+++ Server/controller/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../model/section");
 const Course = require("../model/Course");
+const SubSection = require("../model/subSection")
 
 
 // Create Section
@@ -30,7 +31,7 @@ exports.createSection = async(req,res) => {
                                                     .populate ({
                                                         path : "courseContent",
                                                         populate : {
-                                                          path : "Section"  
+                                                          path : "subsections"  
                                                         },
                                                     })
                                                     .exec()
@@ -59,7 +60,7 @@ exports.createSection = async(req,res) => {
 exports.updateSection = async(req,res) => {
     try{
             //data fetch
-            const {sectionName , sectionId} = req.body
+            const {sectionName , sectionId , courseId} = req.body
             // Validation
             if(!sectionName ||  !sectionId){
                 
@@ -75,11 +76,21 @@ exports.updateSection = async(req,res) => {
                                                              {sectionName},
                                                              {new:true}
              );
+
+             // UPdate The Course With Section
+             const course = await Section.findByIdAndUpdate(courseId).
+                                                            populate({
+                                                                path : "courseContent",
+                                                                populate :{
+                                                                    path: "subsections"
+                                                                },
+                                                            })
+                                                            .exec();
              // Send Response
              return res.status(200).json({
                 success : true,
-                message : "Section Updated SuccessFully ",
-                updateSectionDetails
+                message : `"Section Updated SuccessFully " ${updateSectionDetails}`,
+                data : course
             })
 
      }
@@ -101,17 +112,48 @@ exports.updateSection = async(req,res) => {
 exports.deleteSection = async(req , res) => {
     try{
          // feTCH Data
-         const{sectionId}  = req.body;
+         const{sectionId ,courseId}  = req.body;
+
+         await Course.findByIdAndUpdate(courseId , 
+            {
+                $pull : {
+                    courseContent : sectionId
+                }
+            }
+         )
+
+         const section = await Section.findById(sectionId)
+         console.log(sectionId , courseId);
+
+         if(!section) {
+            return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+         }
+
+         	//delete sub section
+		    await SubSection.deleteMany({_id: {$in: section.subsections}});
+
+         
 
          // find And Delete 
          const deleteSection = await Section.findByIdAndDelete(sectionId);
          // update thre course with updated section 
          
+         const course = await Course.findById(courseId).populate({
+            path : "courseContent",
+            populate : {
+                path: "subsections"
+            }
+         })
+         .exec()
+         
          // Send Response
          return res.status(200).json({
             success : true,
             message : "Section Deleted SuccessFully ",
-            // updateSectionDetails
+            data : course
         })
 
     }
